Don’t give special treatment to HEAT.NET and PXO
The main motivation behind this change is to make it easier to make a
future commit. That future commit will replace the while loop that
populates the “Choose Connection Type”. It’s going to be easier to
modify that while loop if I don’t have to worry about making sure that
HEAT.NET is listed first and that PXO is listed second.